Separate buffer generation into different utils
