+++ experimental/tools/unit-generators/smv/generators/handshake/mux.py
@@ -1,4 +1,4 @@
-from generators.handshake.buffer import generate_buffer
+from generators.support.tehb import generate_tehb
 from generators.support.utils import *
 
 
@@ -31,7 +31,7 @@ MODULE {name}({", ".join([f"ins_valid_{n}" for n in range(size)])}, index, index
   index_ready := !index_valid | tehb_ins_valid & inner_tehb.ins_ready;
   outs_valid := inner_tehb.outs_valid;
 
-{generate_buffer(f"{name}__tehb_dataless", TEHB_BUFFER_PARAMS(HANDSHAKE_CONTROL_TYPE))}
+{generate_tehb(f"{name}__tehb_dataless", {ATTR_DATA_TYPE: HANDSHAKE_CONTROL_TYPE.mlir_type})}
 """
 
 
@@ -58,5 +58,5 @@ MODULE {name}({", ".join([f"ins_{n}, ins_valid_{n}" for n in range(size)])}, ind
   outs_valid := inner_tehb.outs_valid;
   outs := inner_tehb.outs;
 
-{generate_buffer(f"{name}__tehb", TEHB_BUFFER_PARAMS(data_type))}
+{generate_tehb(f"{name}__tehb_dataless", {ATTR_DATA_TYPE: data_type.mlir_type})}
 """
