+++ experimental/tools/unit-generators/smv/generators/support/utils.py
@@ -66,10 +66,3 @@ class SmvScalarType:
 
 
 HANDSHAKE_CONTROL_TYPE = SmvScalarType("!handshake.control<>")
-
-
-def TEHB_BUFFER_PARAMS(data_type):
-  return {ATTR_SLOTS: 1, ATTR_TIMING: "R: 1", ATTR_PORT_TYPES: {"outs": data_type.mlir_type}}
-
-def OEHB_BUFFER_PARAMS(data_type):
-  return {ATTR_SLOTS: 1, ATTR_TIMING: "R: 0", ATTR_PORT_TYPES: {"outs": data_type.mlir_type}}
